feat: move sign up logic into auth controller
