+++ src/main/java/com/develcode/develfood/controller/AuthenticationController.java
@@ -1,9 +1,14 @@
 package com.develcode.develfood.controller;
 
 import com.develcode.develfood.dto.AuthData;
+import com.develcode.develfood.dto.CustomerDataDto;
+import com.develcode.develfood.dto.CustomerSignUpDto;
 import com.develcode.develfood.dto.JwtTokenData;
-import com.develcode.develfood.infra.security.TokenService;
+import com.develcode.develfood.dto.RestaurantDataDto;
+import com.develcode.develfood.dto.RestaurantSignUpDto;
+import com.develcode.develfood.infra.security.JwtService;
 import com.develcode.develfood.model.User;
+import com.develcode.develfood.services.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -16,20 +21,36 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/login")
+@RequestMapping("/api/auth")
 public class AuthenticationController {
 
     private final AuthenticationManager authenticationManager;
 
-    private final TokenService tokenService;
+    private final JwtService jwtService;
 
-    @PostMapping
+    private final UserService userService;
+
+    @PostMapping("/login")
     public ResponseEntity<JwtTokenData> login(@RequestBody @Valid AuthData authData) {
         var authenticationToken = new UsernamePasswordAuthenticationToken(authData.getEmail(), authData.getPassword());
         var authentication = authenticationManager.authenticate(authenticationToken);
 
-        var tokenJwt = tokenService.generateToken( (User) authentication.getPrincipal());
+        var tokenJwt = jwtService.generateToken( (User) authentication.getPrincipal());
 
         return ResponseEntity.ok(new JwtTokenData(tokenJwt));
+   }
+
+    @PostMapping("/cliente/cadastro")
+    public ResponseEntity<CustomerDataDto> signUp(@RequestBody @Valid CustomerSignUpDto customerSignUpDto) {
+        var newUser = userService.newCustomerSignUp(customerSignUpDto);
+
+        return ResponseEntity.ok(newUser);
+    }
+
+    @PostMapping("/restaurante/cadastro")
+    public ResponseEntity<RestaurantDataDto> signUp(@RequestBody @Valid RestaurantSignUpDto restaurantSignUpDto) {
+        var newRestaurant = userService.newRestaurantSignUp(restaurantSignUpDto);
+
+        return ResponseEntity.ok(newRestaurant);
     }
 }
