+++ src/main/java/com/develcode/develfood/controller/RestaurantController.java
@@ -1,30 +1,22 @@
 package com.develcode.develfood.controller;
 
 
-import com.develcode.develfood.dto.CustomerDataDto;
-import com.develcode.develfood.dto.CustomerSignUpDto;
-import com.develcode.develfood.dto.RestaurantDataDto;
-import com.develcode.develfood.dto.RestaurantSignUpDto;
-import com.develcode.develfood.services.UserService;
-import jakarta.validation.Valid;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
+import com.develcode.develfood.services.RestaurantService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/restaurantes")
 public class RestaurantController {
 
-    @Autowired
-    private UserService userService;
+    private final RestaurantService restaurantService;
 
-    @PostMapping("/cadastro")
-    public ResponseEntity<RestaurantDataDto> signUp(@RequestBody @Valid RestaurantSignUpDto restaurantSignUpDto) {
-        var newRestaurant = userService.newRestaurantSignUp(restaurantSignUpDto);
-
-        return ResponseEntity.ok(newRestaurant);
-    }
+//    @GetMapping
+//    public ResponseEntity<Page<RestaurantListDataDto>> list(Pageable pageable) {
+//        var page = restaurantService.findAllByActive(pageable).map(RestaurantListDataDto::new);
+//
+//        return ResponseEntity.ok(page);
+//    }
 }
